feat(frontend): add loading spinning gif
Notes: Spinning gif code is added and is working as expected.
However, it does not spin has browser is heavily processing
something. I want to address this if I can to make the application
more smooth. This may or may not be a relatively large refactoring
so this work will be done in its own branch

Closes #42